fix(lint): resolve style/security findings (gosec/gocyclo/revive/usestdlibvars/nakedret/gocritic/whitespace/unconvert)
- gosec: justified nolint for protocol md5/sha1 checksums; fixed context-cancel
  leak in blob_store_migration_service; tightened blob dir perms to 0o750;
  justified content-addressed path open and detached-context goroutines.
- usestdlibvars/gocritic/nakedret/unconvert/whitespace: mechanical fixes.
- revive: curated rule subset (exported-doc/unused-param/var-naming deferred).

+++ internal/metrics/metrics_test.go
@@ -43,6 +43,6 @@ func TestCounters_Requests(t *testing.T) {
 	before := metrics.RequestsTotal.Load()
 	metrics.RequestsTotal.Add(3)
 	snap := metrics.Snapshot()
-	assert.Equal(t, int64(before+3), snap["requests_total"])
+	assert.Equal(t, before+3, snap["requests_total"])
 	metrics.RequestsTotal.Add(-3)
 }
